chore: 해당 클래스들은 타이틀뷰에서만 사용하기 때문에 파일을 하나로 합침
