--- iOS/Todo/Todo/Views/PlusButton.swift
@@ -1,26 +1,0 @@
-//
-//  PlusButton.swift
-//  Todo
-//
-//  Created by kimdo2297 on 2020/04/07.
-//  Copyright © 2020 Jason. All rights reserved.
-//
-
-import UIKit
-
-final class PlusButton: UIButton {
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        configure()
-    }
-    
-    required init?(coder: NSCoder) {
-        super.init(coder: coder)
-        configure()
-    }
-    
-    private func configure() {
-        translatesAutoresizingMaskIntoConstraints = false
-        setImage(UIImage(systemName: "plus"), for: .normal)
-    }
-}
